🎨 히스토리 관련 함수 수정
- parseHistory 위치 변경
- getUser
