+++ client/src/components/molecules/HistoryLog/index.tsx
@@ -1,7 +1,8 @@
 import { Title } from 'components/atoms';
 import { useRecoilValue } from 'recoil';
-import { queryUserListState } from 'recoil/user-list';
+import { userListState } from 'recoil/project';
 import { IHistoryData } from 'types/history';
+import { getUser } from 'utils/helpers';
 import { Profile } from '..';
 import { Wrapper } from './style';
 
@@ -10,13 +11,13 @@ interface IProps {
 }
 
 const HistoryLog: React.FC<IProps> = ({ history }) => {
-  const userList = useRecoilValue(queryUserListState);
+  const userList = useRecoilValue(userListState);
 
   return (
     <Wrapper>
       {history && userList ? (
         <>
-          <Profile user={userList[history.user]} />
+          <Profile user={getUser(history.user, userList)!} />
           <Title titleStyle='large' color='white' margin='0 0 0 1rem' lineHeight={2.5}>
             {history.type}
           </Title>
