+++ client/src/utils/helpers.ts
@@ -1,5 +1,6 @@
 import { THistoryEventData } from 'types/event';
 import { IMindNodes } from 'types/mindmap';
+import { IUser } from 'types/user';
 
 interface ICoord {
   x: number;
@@ -114,8 +115,10 @@ const setTreeLevel = (mindNodes: IMindNodes, nodeId: number, depth: number) => {
   node.level = idxToLevel(depth);
   node.children.forEach((childId) => setTreeLevel(mindNodes, childId, depth + 1));
 };
+const getUser = (userId: string | undefined, userList: IUser[]) => userList.find((user) => user.id === userId);
 
 export {
+  getUser,
   getChildLevel,
   fillPayload,
   getNewNode,
